XmlPoke - Extended the UpdateOrInsertValueInFile with "shouldBeFirst" parameter. If it's true, the new element will be inserted as the first child. Doesn't move existing elements though. Reason: "configSections" element must be the first in a web.config... Test: Xml/XmlPokeTaskSpecs.cs -> "when_inserting_into_the_settings_in_an_xml_file_using_shouldBeFirst"
note: API at XmlPokeTask level is pretty ugly, because replaced a Dictionary<string, string> with a Dictionary<string, <Tuple<string, bool>>, but that does not affect the fluent API, since adding a new overload to XmlPokeOptions.SetOrInsert covers the internals.

Modified the when_adding_Elmah_to_web_config_file test, because it updated all <location> parts in web.config, even if it was present in the source for a completely different pat. Also configSections is set to be first.